FLIX-208: Create view list requests api for repairer

+++ src/main/java/com/fu/flix/entity/Voucher.java
@@ -24,7 +24,7 @@ public class Voucher {
 
     private String description;
 
-    private Double minOrderPrice;
+    private Long minOrderPrice;
 
     private Long remainQuantity;
